+++ map-backend/src/main/java/pl/pwr/bdmap/dto/NodeDTO.java
@@ -4,8 +4,8 @@ import java.sql.Timestamp;
 
 public record NodeDTO(
         int id,
-        int posX,
-        int posY,
+        double posX,
+        double posY,
         boolean isBlocked,
         Timestamp timestamp,
         String nodeType
